% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{error_bar}
\alias{error_bar}
\title{error_bar}
\usage{
error_bar(x, y, upper, lower = upper, width = 0.1, ...)
}
\arguments{
\item{x}{X-position on the plot.}

\item{y}{Y-position on the plot.}

\item{upper}{Size of the upper error bar.}

\item{lower}{Size of the lower error bar. By default, it equals the upper error bar.}

\item{width}{Width of the erro bar whisker.}

\item{...}{Pass any other parameter of the corresponding arrows function (most of them should work).}
}
\description{
Put error bars on top of your bar plots. This functionality is now integrated into MarkdownReporter's wbarplot() function
}
\examples{
error_bar (x =  , y =  , upper =  , lower = upper, width  = 0.1, ... =  )
}
