% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{translate}
\alias{translate}
\title{translate}
\usage{
translate(vec, oldvalues, newvalues)
}
\arguments{
\item{vec}{set of values where you want to replace}

\item{oldvalues}{oldvalues (from)}

\item{newvalues}{newvalues (to)}
}
\description{
Replaces a set of values in a vector with another set of values, it translates your vector. Oldvalues and newvalues have to be 1-to-1 corespoding vectors.
}
\examples{
A=1:3; translate(vec = A, oldvalues =2:3 , newvalues = letters[1:2])
}
